+++ app/components/Dashboard/Dashboard.jsx
@@ -1,14 +1,10 @@
 import React from "react";
-import Immutable from "immutable";
-import DashboardList from "./DashboardList";
-import { RecentTransactions } from "../Account/RecentTransactions";
 import Translate from "react-translate-component";
 import MarketCard from "./MarketCard";
 import utils from "common/utils";
 import { Apis } from "bitsharesjs-ws";
 import LoadingIndicator from "../LoadingIndicator";
 import LoginSelector from "../LoginSelector";
-import cnames from "classnames";
 import SettingsActions from "actions/SettingsActions";
 import SettingsStore from "stores/SettingsStore";
 import { connect } from "alt-react";
@@ -40,6 +36,7 @@ class Dashboard extends React.Component {
                 ["CNY", "GDEX.ETH"],
                 ["CNY", "GDEX.EOS"],
                 ["CNY", "GDEX.BTO"],
+                ["CNY", "GDEX.BTM"],
                 ["OPEN.BTC", "BTS"],
                 ["OPEN.BTC", "OPEN.ETH"],
                 ["OPEN.BTC", "OPEN.DASH"],
@@ -84,16 +81,13 @@ class Dashboard extends React.Component {
 
         this.state = {
             width: null,
-            showIgnored: false,
             featuredMarkets: marketsByChain[chainID] || marketsByChain["4018d784"],
             newAssets: [
 
-            ],
-            currentEntry: props.currentEntry
+            ]
         };
 
         this._setDimensions = this._setDimensions.bind(this);
-        // this._sortMarketsByVolume = this._sortMarketsByVolume.bind(this);
     }
 
     componentDidMount() {
@@ -108,13 +102,9 @@ class Dashboard extends React.Component {
             !utils.are_equal_shallow(nextProps.lowVolumeMarkets, this.props.lowVolumeMarkets) ||
             !utils.are_equal_shallow(nextState.newAssets, this.state.newAssets) ||
             nextProps.linkedAccounts !== this.props.linkedAccounts ||
-            // nextProps.marketStats !== this.props.marketStats ||
-            nextProps.ignoredAccounts !== this.props.ignoredAccounts ||
             nextProps.passwordAccount !== this.props.passwordAccount ||
             nextState.width !== this.state.width ||
-            nextProps.accountsReady !== this.props.accountsReady ||
-            nextState.showIgnored !== this.state.showIgnored ||
-            nextState.currentEntry !== this.state.currentEntry
+            nextProps.accountsReady !== this.props.accountsReady
         );
     }
 
@@ -130,31 +120,15 @@ class Dashboard extends React.Component {
         }
     }
 
-    _onToggleIgnored() {
-        this.setState({
-            showIgnored: !this.state.showIgnored
-        });
-    }
-
-    _onSwitchType(type) {
-        this.setState({
-            currentEntry: type
-        });
-        SettingsActions.changeViewSetting({
-            dashboardEntry: type
-        });
-    }
-
     render() {
         let { linkedAccounts, myIgnoredAccounts, accountsReady, passwordAccount } = this.props;
-        let {width, showIgnored, featuredMarkets, newAssets, currentEntry} = this.state;
+        let { featuredMarkets, newAssets} = this.state;
 
         if (passwordAccount && !linkedAccounts.has(passwordAccount)) {
             linkedAccounts = linkedAccounts.add(passwordAccount);
         }
         let names = linkedAccounts.toArray().sort();
         if (passwordAccount && names.indexOf(passwordAccount) === -1) names.push(passwordAccount);
-        let ignored = myIgnoredAccounts.toArray().sort();
 
         let accountCount = linkedAccounts.size + myIgnoredAccounts.size + (passwordAccount ? 1 : 0);
 
@@ -169,10 +143,12 @@ class Dashboard extends React.Component {
             let isLowVolume = this.props.lowVolumeMarkets.get(pair[1] + "_" + pair[0]) || this.props.lowVolumeMarkets.get(pair[0] + "_" + pair[1]);
             if (!isLowVolume) validMarkets++;
             let className = "";
-            if (validMarkets > 9) {
-                className += ` show-for-${!accountCount ? "xlarge" : "large"}`;
+            if (validMarkets > 24) {
+                return null;
+            } else if (validMarkets > 15) {
+                className += " show-for-large";
             } else if (validMarkets > 6) {
-                className += ` show-for-${!accountCount ? "large" : "medium"}`;
+                className += " show-for-medium";
             }
 
             return (
@@ -185,7 +161,6 @@ class Dashboard extends React.Component {
                     base={pair[1]}
                     invert={pair[2]}
                     isLowVolume={isLowVolume}
-                    hide={validMarkets > 20}
                 />
             );
         }).filter(a => !!a);
@@ -193,9 +168,6 @@ class Dashboard extends React.Component {
         if (!accountCount) {
             return <LoginSelector />;
         }
-
-        const entries = ["accounts", "contacts", "recent"];
-        const activeIndex = entries.indexOf(currentEntry);
 
         return (
             <div ref="wrapper" className="grid-block page-layout vertical">
@@ -206,44 +178,6 @@ class Dashboard extends React.Component {
                     {this.props.onlyAccounts ? null : <div className="grid-block small-up-1 medium-up-3 large-up-4 no-overflow fm-outer-container">
                         {markets}
                     </div>}
-
-                    {accountCount ? (
-                        <div style={{paddingBottom: "3rem"}}>
-                            <div className="hide-selector" style={{paddingBottom: "1rem"}}>
-                                {entries.map((type, index) => {
-                                    return (
-                                        <div key={type} className={cnames("inline-block", {inactive: activeIndex !== index})} onClick={this._onSwitchType.bind(this, type)}>
-                                            <Translate content={`account.${type}`} />
-                                        </div>
-                                    );
-                                })}
-                            </div>
-
-                            {(currentEntry === "accounts" || currentEntry === "contacts") ? <div className="generic-bordered-box" style={{marginBottom: 5}}>
-                                <div className="box-content">
-                                    <DashboardList
-                                        accounts={Immutable.List(names)}
-                                        ignoredAccounts={Immutable.List(ignored)}
-                                        width={width}
-                                        onToggleIgnored={this._onToggleIgnored.bind(this)}
-                                        showIgnored={showIgnored}
-                                        showMyAccounts={currentEntry === "accounts"}
-                                    />
-                                    {/* {showIgnored ? <DashboardList accounts={Immutable.List(ignored)} width={width} /> : null} */}
-                                </div>
-                            </div> : null}
-
-                            {currentEntry === "recent" ? <RecentTransactions
-                                style={{marginBottom: 20, marginTop: 20}}
-                                accountsList={linkedAccounts}
-                                limit={10}
-                                compactView={false}
-                                fullHeight={true}
-                                showFilters={true}
-                                dashboard
-                            /> : null}
-                        </div>
-                    ) : null}
                 </div>
             </div>
         );