+++ app/components/Exchange/Exchange.jsx
@@ -21,14 +21,13 @@ import cnames from "classnames";
 import market_utils from "common/market_utils";
 import {Asset, Price, LimitOrderCreate} from "common/MarketClasses";
 import ConfirmOrderModal from "./ConfirmOrderModal";
-// import IndicatorModal from "./IndicatorModal";
-import OpenSettleOrders from "./OpenSettleOrders";
 import Highcharts from "highcharts/highstock";
 import ExchangeHeader from "./ExchangeHeader";
 import Translate from "react-translate-component";
 import { Apis } from "bitsharesjs-ws";
 import GatewayActions from "actions/GatewayActions";
 import { checkFeeStatusAsync } from "common/trxHelper";
+import LoadingIndicator from "../LoadingIndicator";
 
 Highcharts.setOptions({
     global: {
@@ -992,7 +991,7 @@ class Exchange extends React.Component {
         }
 
         let orderMultiplier = leftOrderBook ? 2 : 1;
-        const minChartHeight = 300
+        const minChartHeight = 300;
         const height = Math.max(
             this.state.height > 1100 ? chartHeight : chartHeight - 125,
             minChartHeight
@@ -1114,7 +1113,22 @@ class Exchange extends React.Component {
             />
         );
 
-        return (
+        return (<div className="grid-block vertical">
+            {!this.props.marketReady ? <LoadingIndicator /> : null}
+                    <ExchangeHeader
+                        account={this.props.currentAccount}
+                        quoteAsset={quoteAsset} baseAsset={baseAsset}
+                        hasPrediction={hasPrediction} starredMarkets={starredMarkets}
+                        lowestAsk={lowestAsk} highestBid={highestBid}
+                        lowestCallPrice={lowestCallPrice}
+                        showCallLimit={showCallLimit} feedPrice={feedPrice}
+                        marketReady={marketReady} latestPrice={latestPrice}
+                        showDepthChart={showDepthChart}
+                        onSelectIndicators={this._onSelectIndicators.bind(this)}
+                        marketStats={marketStats}
+                        onToggleCharts={this._toggleCharts.bind(this)}
+                        showVolumeChart={showVolumeChart}
+                    />
             <div className="grid-block page-layout market-layout">
                     <AccountNotifications/>
                     {/* Main vertical block with content */}
@@ -1127,22 +1141,6 @@ class Exchange extends React.Component {
 
                     {/* Center Column */}
                     <div style={{paddingTop: 0}} className={cnames("grid-block main-content vertical no-overflow")} >
-
-                        {/* Top bar with info */}
-                        <ExchangeHeader
-                            quoteAsset={quoteAsset} baseAsset={baseAsset}
-                            hasPrediction={hasPrediction} starredMarkets={starredMarkets}
-                            lowestAsk={lowestAsk} highestBid={highestBid}
-                            lowestCallPrice={lowestCallPrice}
-                            showCallLimit={showCallLimit} feedPrice={feedPrice}
-                            marketReady={marketReady} latestPrice={latestPrice}
-                            showDepthChart={showDepthChart}
-                            onSelectIndicators={this._onSelectIndicators.bind(this)}
-                            marketStats={marketStats}
-                            onToggleCharts={this._toggleCharts.bind(this)}
-                            showVolumeChart={showVolumeChart}
-                        />
-
                         <div className="grid-block vertical no-padding ps-container" id="CenterContent" ref="center">
                         {!showDepthChart ? (
                             <div className="grid-block shrink no-overflow" id="market-charts" >
@@ -1228,7 +1226,7 @@ class Exchange extends React.Component {
                                 {isFrozen ? <div className="error small-12 no-overflow" style={{margin: "0 10px", lineHeight: "1.2rem"}}><Translate content="exchange.market_frozen" asset={frozenAsset} component="p"/></div> : null}
                                 {buyForm}
                                 {sellForm}
-
+                                
                                 <MarketHistory
                                     className={cnames(
                                         !smallScreen && !leftOrderBook ? "medium-6 xlarge-4" : "",
@@ -1273,36 +1271,19 @@ class Exchange extends React.Component {
                                     )}
                                     key="open_orders"
                                     orders={marketLimitOrders}
+                                    settleOrders={marketSettleOrders}
                                     currentAccount={currentAccount}
                                     base={base}
                                     quote={quote}
                                     baseSymbol={baseSymbol}
                                     quoteSymbol={quoteSymbol}
+                                    activeTab={this.props.viewSettings.get("ordersTab")}
                                     onCancel={this._cancelLimitOrder.bind(this)}
                                     flipMyOrders={this.props.viewSettings.get("flipMyOrders")}
                                     feedPrice={this.props.feedPrice}
                                 />) : null}
                             </div>
-
-
-                            {/* Settle Orders */}
-
-                            {(base.get("id") === "1.3.0" || quote.get("id") === "1.3.0") ? (
-                            <OpenSettleOrders
-                                key="settle_orders"
-                                className={cnames(!smallScreen && !leftOrderBook ? "medium-6 xlarge-4 order-12" : "",
-                                    `small-12 medium-6 no-padding align-spaced ps-container middle-content order-12`
-                                )}
-                                orders={marketSettleOrders}
-                                base={base}
-                                quote={quote}
-                                baseSymbol={baseSymbol}
-                                quoteSymbol={quoteSymbol}
-                            />) : null}
-
-
                         </div>{ /* end CenterContent */}
-
 
                     </div>{/* End of Main Content Column */}
 
@@ -1378,6 +1359,7 @@ class Exchange extends React.Component {
                             account={currentAccount}
                         /> : null}
                 {/* End of Second Vertical Block */}
+                </div>
                 </div>
         );
     }