+++ app/components/Modal/DepositModal.jsx
@@ -13,7 +13,7 @@ import QRCode from "qrcode.react";
 import DepositWithdrawAssetSelector from "../DepositWithdraw/DepositWithdrawAssetSelector.js";
 
 class DepositModalContent extends DecimalChecker {
-    
+
     constructor() {
         super();
 
@@ -49,6 +49,8 @@ class DepositModalContent extends DecimalChecker {
             let assetName = backedAsset[1];
             let assetGateway = backedAsset[0];
             this._getDepositAddress(assetName, assetGateway);
+        } else {
+            this.setState({ selectedAsset: "BTS" });
         }
     }
 
@@ -150,7 +152,7 @@ class DepositModalContent extends DecimalChecker {
                 fetchingAddress: false
             });
         } else {
-            console.log("Withdraw Modal Error: Unknown Gateway " + selectedGateway + "(asset: " + selectedAsset + ")");
+            console.log("Withdraw Modal Error: Unknown Gateway " + selectedGateway + " for asset " + selectedAsset);
         }
 
         this.setState({
@@ -187,7 +189,6 @@ class DepositModalContent extends DecimalChecker {
     render() {
         let {selectedAsset, selectedGateway, depositAddress, fetchingAddress, gatewayStatus, backingAsset} = this.state;
         let {account} = this.props;
-
         let usingGateway = true;
 
         if(selectedGateway == null && selectedAsset == "BTS") {
@@ -215,6 +216,8 @@ class DepositModalContent extends DecimalChecker {
 
         const logo = require("assets/logo-ico-blue.png");
 
+        //console.log(selectedAsset + " w/ " + selectedGateway + " ", backingAsset, depositAddress);
+
         return (
             <div className="DepositModal">
                 <div className="canvas grid-block vertical no-overflow">
@@ -233,7 +236,7 @@ class DepositModalContent extends DecimalChecker {
                                 <div className="inline-label input-wrapper">
                                     <DepositWithdrawAssetSelector
                                         defaultValue={selectedAsset}
-                                        onSelect={this.onAssetSelected.bind(this)} 
+                                        onSelect={this.onAssetSelected.bind(this)}
                                         selectOnBlur />
                                 </div>
                             </div>
@@ -243,7 +246,7 @@ class DepositModalContent extends DecimalChecker {
                                 <div className="no-margin no-padding">
                                     <section className="block-list">
                                         <label className="left-label"><Translate content="modal.deposit.gateway" />
-                                            {selectedGateway ? <span>&nbsp;<Icon name="question-circle" onClick={this._openGatewaySite.bind(this)}/></span> : null}
+                                            {selectedGateway ? <span style={{cursor: "pointer"}}>&nbsp;<Icon name="question-circle" onClick={this._openGatewaySite.bind(this)}/></span> : null}
                                             <span className="floatRight error-msg">
                                                 {selectedGateway && !gatewayStatus[selectedGateway].enabled ? <Translate content="modal.deposit.disabled" /> : null}
                                                 {depositAddress && depositAddress.error ? <Translate content="modal.deposit.wallet_error" /> : null}
@@ -262,10 +265,10 @@ class DepositModalContent extends DecimalChecker {
                                     </section>
                                 </div>
                             </div> : null}
-                        {!fetchingAddress ? 
+                        {!fetchingAddress ?
                             (!usingGateway || (usingGateway && selectedGateway && gatewayStatus[selectedGateway].enabled)) && depositAddress && !depositAddress.memo ?
-                                <div className="container-row" style={{textAlign: "center"}}>{QR}</div> : 
-                                null : 
+                                <div className="container-row" style={{textAlign: "center"}}>{QR}</div> :
+                                null :
                             <div className="container-row" style={{textAlign: "center"}}><LoadingIndicator type="three-bounce" /></div>
                         }
                         {selectedGateway && gatewayStatus[selectedGateway].enabled && depositAddress && !depositAddress.error ?
@@ -303,7 +306,7 @@ class DepositModalContent extends DecimalChecker {
                         : null}
                         {!usingGateway ?
                             <div className="container-row deposit-directly">
-                                <p>{account}</p>
+                                <p><span className="send-name">{account}</span></p>
                                 <p>
                                     <Translate content="modal.deposit.bts_transfer_description" />
                                 </p>